Add Butterfinger, change the API (#3)
This was intended to add the Butterfinger error type, which it does. It
also completely changes how ErrorMechanism and ErrorType is configured.

The idea is that the abstract base classes ErrorMechanism and ErrorType
become stateful, and that they are configured upon initialization.

I made this change because I figured that ErrorTypes need to be
configured with configurations that are very different. Having a
dataclass that stores that configuration allows us to control all
parameters in one place, which is handy when adding more ErrorTypes in
the future I believe.

This does change however how error_generation can be used, check the
samples.ipynb for reference.