Apply error mechanisms per column, not globally
Until this point, error mechanisms return an error mask that has the
same dimensions as the entire dataframe. This is now improved, where
each mechanism is applied to one column.

This allows in the long run to have complex scenarios, in which we
iteratively have error mechanisms build the error mask - that will be
handy for the mid-level API.

The commit changes the API a little bit.