Make Outlier respect dtype, remove Clipping
I added support for `Outlier` to respect dtypes - series that contain
integers won't be cast to float, but preserve the int dtype.

I ended up removing the `Clipping` Error Type, because it introduces
a dependency between error_pct and the error type that the library
currently does not support. A bigger effort would be needed to
properly support `Clipping`.

I added one example of how to use `Outlier`.