Feat/typo one word (#21)
I observed that many datasets contain empty strings in columns into
which I want to insert Typos. Currently, this causes an error. With this
PR, an empty string `""` will be replaced by a random character, e.g.
`"a"`. While that's not a perfect solution, it gets us going.

In the future, we could look into dependencies between error types and
error mechanisms (e.g., insert Typos into values that are not empty
strings, yet preserve the error mechanism as good as possible).

**Changes**
- Refactor tests, move them into separate folders and introduce
`conftest.py` for shared fixtures.
- Add a (very basic) unit test for `Typo`
- Make `Typo` work with empty strings

---------

Co-authored-by: Sebastian Jäger <github@sebastian-jaeger.me>