Fix Issue where High-Level API using ECAR with Given Seed Generated all Errors in the Same Columns (#23)
The proposed fix to the above mentioned issue uses the given seed to
setup a random generator, which is used to sample seeds for `ErrorType`s
and `ErrorMechanism`s. This way, everything is still deterministic but
the errors are distributed over the entire table and not clumped
together in certain rows.

---------

Signed-off-by: Sebastian Jäger <git@sebastian-jaeger.me>