chore: CI after merging into main, add pyton 3.13
CI is currently only run on PRs. However, it is a good practice to also
run CI on `main` after merging to make sure that checks on `main` always
pass.

I also added Python 3.13 to the `python-version`s in the workflows - I
don't see a reason why we wouldn't support 3.13.