Polish README
This adds shields to the README that show supported python versions and
the status of our github workflows. The shields are currently grayed
out, but once this PR is merged, they will be show how the workflows on
the `main` branches did.

I gave the README another sturcture, it now starts with a tl;dr section
at the very top.

I also added a minimal example right to the README - I find that nice, a
lot of people won't click any links to example notebooks.

I also added a section for citing the paper.